feat(builder): change behavior of create2 to fail on existing contract
also do some refactoring wrt handling of `PackageReference` with `ChainDefinition`, which lead to a bit of a rabbit hole

fixes https://linear.app/usecannon/issue/CAN-136/change-behaviour-of-contract-cannon-build-with-create2-to-fail-on
